+++ tests/test_mcp.py
@@ -793,7 +793,7 @@ EXPECTED_TOOL_PARAMETERS = {
     "debug_info": ("app_id env_file", "app_id"),
     "delete_app": ("app_id env_file", "app_id"),
     "deploy": (
-        "app_id branch env_file project_dir project_id version",
+        "app_id branch env_file project_dir project_id server_start_timeout version",
         "app_id project_id",
     ),
     "ensure_app": (